Move is_overlong to helpers
As I started to move violations to the individual rule files I realized
I made a mistake making this a rule file, as it's not a rule, just a
helper for other rules. This fixes that by moving it to helpers.
