+++ src/rules/pycodestyle/rules/mod.rs
@@ -5,7 +5,6 @@ pub use do_not_assign_lambda::do_not_assign_lambda;
 pub use do_not_use_bare_except::do_not_use_bare_except;
 pub use doc_line_too_long::doc_line_too_long;
 pub use invalid_escape_sequence::invalid_escape_sequence;
-pub use is_overlong::is_overlong;
 pub use line_too_long::line_too_long;
 pub use literal_comparisons::literal_comparisons;
 pub use mixed_spaces_and_tabs::mixed_spaces_and_tabs;
@@ -20,7 +19,6 @@ mod do_not_assign_lambda;
 mod do_not_use_bare_except;
 mod doc_line_too_long;
 mod invalid_escape_sequence;
-mod is_overlong;
 mod line_too_long;
 mod literal_comparisons;
 mod mixed_spaces_and_tabs;
